Fix settings race condition when using broadcast
This commit fixes a race condition between setting up parameter and sending data via USB.
The problem was very apparent when using Broadcast as it made the radio send packet on the wrong address.
It could also be apparent when using the radio with a Swarm.

After this commit, Crazyradio 2 should work well with ROS.