Fix (no)ACK bit for compatiblity with ESB
Full ESB PRX implementations like the nRF24's and Nordic's ESB implementation
reads a "ACK" bit in the packet to know if an ack needs to be sent or not.

This ACK bit needs to be set to 1 in order for the PRX device to send an ack (assuming dynamic ack is enabled).

This commit hardcode the ACK bit to 1 in the S1 field of the packet to work with such implementations.